Make another header file self-sufficiency change
that the recently added test for self-sufficiency did NOT catch.

The issue here is m_random.h depends upon the macro "DA"
but doesn't include a definition of that macro, but the
compiler doesn't catch this issue when just compiling the header
on its own, and I'm not sure why, but not too concerned.