Create libsdldoom library as a libtool library
because when doing this it is simpler* to combine
this library with another library to create an combined
library that acts as if its just one huge library.

This is something we're going to do in a future commit
so we can create a massive single library of all prboom code
which can then be linked against future unit tests.

* - all from my very limited uneducated experience