Fixed a warning
I_CalculateRes was being called with signed int args, but defined
with unsigned int parameters. Thankfully, I_CalculateRes doesn't
actually benefit from having unsigned int parameters, so there's
no loss in using signed parameters instead in order to be
consistent with the usage of this function, and avoid this warning.