Fix memory leak
that happens when a default_t setting of type `def_str` that
has both a non-null default and the flag S_CHOICE is edited by
the user (an example of such a setting is "gl_tex_filter_string")

When this happens the previously set and owned default char*
value is just lost to the ether, to never be reclaimed by memory
management.  This change fixes that.

Also, the code as it was would cause a crash if there ever was
a def_str setting that was associated with a setupscreen other
than ss_none and the user decides to reset that setting to its
default value (this would cause a statically allocated (i.e
non-Zone allcoated) char* to be deallocated by the zone
allocator, causing a process error).  This change also fixes that
issue.  Note that there currently isn't any default_t setting
that has all these properties and exhibits this crash issue, but
there's nothing stopping such a setting from being created.