Silence a [-Wcast-qual] warning
by using GCC diagnostics

The way the code is designed intends for the char* values
associated with a default_t setting with a type of def_str
to be both free'able and const.  The previous commit now
makes it consistently true that all such values are free'able,
so we can now safely ignore the const-ness in the case that
we want to free the underlying memory before losing the
only reference we have to that memory.